fix(lottery-v2): Testnet default gas values and round states
