+++ src/state/lottery/index.ts
@@ -4,7 +4,7 @@ import BigNumber from 'bignumber.js'
 import { LotteryState, LotteryRound, LotteryStatus, PastLotteryRound, UserLotteryHistory } from 'state/types'
 import { getLotteryV2Contract } from 'utils/contractHelpers'
 import makeBatchRequest from 'utils/makeBatchRequest'
-import { getPastLotteries, getUserPastLotteries } from './helpers'
+import { getPastLotteries, getUserPastLotteries, fetchLottery, fetchPublicData, fetchTickets } from './helpers'
 
 interface PublicLotteryData {
   currentLotteryId: string
@@ -33,97 +33,6 @@ const initialState: LotteryState = {
   },
   pastLotteries: null,
   userLotteryHistory: null,
-}
-
-const lotteryContract = getLotteryV2Contract()
-
-// TODO: Move SC fetches into helper
-export const fetchLottery = async (lotteryId: string) => {
-  try {
-    const lotteryData = await lotteryContract.methods.lotteries(lotteryId).call()
-    const {
-      status,
-      startTime,
-      endTime,
-      priceTicketInCake,
-      discountDivisor,
-      treasuryFee,
-      firstTicketId,
-      lastTicketId,
-      amountCollectedInCake,
-      finalNumber,
-    } = lotteryData
-    const priceTicketInCakeAsBN = new BigNumber(priceTicketInCake as string)
-    const amountCollectedInCakeAsBN = new BigNumber(amountCollectedInCake as string)
-    const statusKey = Object.keys(LotteryStatus)[status]
-    return {
-      isLoading: false,
-      status: LotteryStatus[statusKey],
-      startTime,
-      endTime,
-      priceTicketInCake: priceTicketInCakeAsBN.toJSON(),
-      discountDivisor,
-      treasuryFee,
-      firstTicketId,
-      lastTicketId,
-      amountCollectedInCake: amountCollectedInCakeAsBN.toJSON(),
-      finalNumber,
-    }
-  } catch (error) {
-    return {
-      isLoading: true,
-      status: LotteryStatus.PENDING,
-      startTime: '',
-      endTime: '',
-      priceTicketInCake: '',
-      discountDivisor: '',
-      treasuryFee: '',
-      firstTicketId: '',
-      lastTicketId: '',
-      amountCollectedInCake: '',
-      finalNumber: '',
-    }
-  }
-}
-
-export const fetchPublicData = async () => {
-  try {
-    const [currentLotteryId, maxNumberTicketsPerBuy] = (await makeBatchRequest([
-      lotteryContract.methods.currentLotteryId().call,
-      lotteryContract.methods.maxNumberTicketsPerBuy().call,
-    ])) as [string, string]
-    return {
-      currentLotteryId,
-      maxNumberTicketsPerBuy,
-    }
-  } catch (error) {
-    return {
-      currentLotteryId: null,
-      maxNumberTicketsPerBuy: null,
-    }
-  }
-}
-
-export const fetchTickets = async (account, lotteryId, cursor) => {
-  try {
-    const userTickets = await lotteryContract.methods
-      .viewUserTicketNumbersAndStatusesForLottery(account, lotteryId, cursor, 1000)
-      .call()
-    const ticketIds = userTickets[0]
-    const ticketNumbers = userTickets[1]
-    const ticketStatuses = userTickets[2]
-
-    const completeTicketData = ticketIds.map((ticketId, index) => {
-      return {
-        id: ticketId,
-        number: ticketNumbers[index],
-        status: ticketStatuses[index],
-      }
-    })
-    return completeTicketData
-  } catch (error) {
-    return null
-  }
 }
 
 export const fetchCurrentLottery = createAsyncThunk<LotteryRound, { currentLotteryId: string }>(
